BR(plugin): Moving getBrand() to the framework.
