+++ zod-plugin/README.md
@@ -14,7 +14,7 @@ This module extends Zod functionality when it's imported:
   - Supports a mapping object or an object transforming function as an argument;
   - Relies on `R.renameKeys()` from the `ramda` library;
 - Alters the `.brand()` method on all Zod schemas:
-  - shorthand for `.meta({ "x-brand": ... })` making the brand available in runtime via `getBrand()` helper;
+  - shorthand for `.meta({ "x-brand": ... })` making the brand available in runtime;
 
 ## Requirements
 
@@ -24,14 +24,8 @@ This module extends Zod functionality when it's imported:
 
 ```ts
 import { z } from "zod";
-import { getBrand } from "@express-zod-api/zod-plugin";
 
 const schema = z.string().example("test").example("another").brand("custom");
 
-getBrand(schema); // "custom"
 schema.meta(); // { examples: ["test", "another"], "x-brand": "custom" }
 ```
-
-## Helpers
-
-- `getBrand()` — retrieves the brand from the schema that was set by its `.brand()` method.
