+++ zod-plugin/src/index.ts
@@ -1,2 +1,1 @@
 import "./runtime"; // side effects here
-export { getBrand } from "./brand";
